Courses Scrapper v0.0.1 CourseAuthor

Provides a set of functions to crawl the data of a course's author. It brings the ability to build a struct with the needed information about the author.

Example

iex> author_tree = <some_nested_tree_object>
<some_nested_tree_object>
iex> CourseAuthor.build_author_struct(author_tree)
%CourseAuthor{
  name: "Rob Percival",
  twitter: "https://twitter.com/techedrob",
  facebook: "https://www.facebook.com/rpcodestars"
  googleplus: "none",
  linkedin: "none",
  youtube: "https://www.youtube.com/user/robpercival"
}

Types

t :: %CourseAuthor{facebook: term, googleplus: String.t, linkedin: String.t, name: String.t, twitter: String.t, youtube: String.t}

Functions

__struct__()

Defines the fixed structure of an author of a course.

build_author_struct(author_tree)

Specs

build_author_struct(author_tree) :: course_author when author_tree: Floki.Finder.html_tree, course_author: CourseAuthor.t

Build a struct that identifies an author of a course based on the HTML document tree passed as argument.

get_author_name(author_tree, query)

Specs

get_author_name(author_tree, query) :: author_name when author_tree: Floki.Finder.html_tree, query: String.t, author_name: String.t

Parses an HTML document tree that represents an author and find the name based on a CSS query string.

get_contact_link(author_tree, keyword)

Specs

get_contact_link(author_tree, keyword) :: contact_link when author_tree: Floki.Finder.html_tree, keyword: String.t, contact_link: String.t

Parses an HTML document tree that represents an author and pattern match against a CSS query string to find his social media contact link.