Courses Scrapper v0.0.1 CoursesScraper.CLI

Defines a command line tool to parse, fetch and process the data from some Udemy courses's sites

Functions

create_output_environment(dir, file)

This function creates the output location where the courses's information will be stored. If the file does exist already, it will simply clear it out for the next write.

handle_file_reading(arg)

Handles the reading process of a file.

This function is resposible for handling the reading process. If it success it returns the string resulting from reading the file.

This function is responsible for handling a posible failure in the reading process. If it fails it sumply prints a message to the console and halts the system with a status of 0.

main(argv)

This is the function that wil be called and run by the escript executable. This function uses the Erlang's :timer.tc function to measure the execution time of the run function defined in this Module. Finally it logs the time alongside some description.

parallel_map_functions(collection, list_of_functions)

This function palallelizes the whole process. It gets a list of paths and a list of functions to be applied to each path. It basically maps a collection into a list of spawned concurrent processes, where each process is responsible for eventually sending a message to the current process with the processed data. That means, that there will be as many processes as paths to be requested in the list of paths. Once all the processes have been spawned, we just wait for the messages to come.

parse_args(argv)

This function reads and parses the arguments passed from the command line and pattern match against the result to decide about the execution flow of the program. We use the Elixir built in parser module (OptionParser) to parse the data and also define some switches to help the user to use the CLI and some aliases to improve the experience of using it.

process(list_of_paths)

This function is responsible for processing the whole list of paths. This transforms step by step the data and finally writes the result into an output file.

read_file_with_paths(path_to_file_with_paths)

Read the file specifed in the path given as argument.

This function is executed when the parser returns a signal indicating help. It prints a message to the console and halts the system with a status of 0.

This function takes the name of the file provided from the CLI and reads it to get the data from it as an string and then splits it into an Elixir List. We perform a sane validation in the middle to ensure that the file was read succesfuly.

run(argv)

This function runs the program based on the arguments passed from the command line. This describes the flow of execution and transformation of the data. First, it parses the arguments from the command line, then it reads the file where the user has indicated the courses paths are located at, and then process those paths to get the courses information

write_data_to_output_file(data)

this function appends the data passed as argument into the output file where the courses's information will be stored.