Courses Scrapper v0.0.1 CoursesScraper.DocumentFetcher

provides a set of high level functions to fetch the data from a website and handle the result into a meaningful value.

Types

body :: String.t

http_resp ::
  {:ok, HTTPoison.Response.t | HTTPoison.AsyncResponse.t} |
  {:error, HTTPoison.Error.t}

info :: body | atom

path :: String.t

response :: {signal, info, path}

signal :: :ok | :error

Functions

build_course_url(course_path)

Specs

build_course_url(path) :: url when url: String.t

Build the full URL to reach the course's website

fetch(course_path)

Specs

fetch(path) :: response

Fetch the data of an Udemy course's website using the HTTPoison library to simulate an HTTP client, and handle the result to return some meaningful data to the caller of the function. This function is embedded into a try do - rescue block because the HTTPoison library has a bug that limits a client to make "too many" requests.