Courses Scrapper v0.0.1 CourseData

Provides a set of functions to crawl the data from a course's HTML document. It brings the ability to build a struct with the needed information about a course.

Example

iex> {:ok, %HTTPoison.Response{status_code: 200, body: body}} = HTTPoison.get("https://www.udemy.com/android-basico/")
{:ok, %HTTPoison.Response{status_code: 200, body: "..."}}
iex> CourseData.extract_data({:ok, body, "android-basico"})
%CourseData{
  url: "https://udemy.com/android-basico",
  topic: "Android básico",
  subcategory: "Mobile Apps",
  source: "Udemy",
  rating: 77,
  price: 20,
  enrolled: 4278,
  category: "Development",
  average_rating: 4.6,
  authors: [
    %CourseAuthor{
      youtube: "https://www.youtube.com/josedlujan1",
      twitter: "https://twitter.com/josedlujan",
      name: "José Luján",
      linkedin: "https://linkedin.com/in/josedlujan/",
      googleplus: "none",
      facebook: "https://www.facebook.com/josedlujan"
    }
  ]
}

Types

authors :: [CourseAuthor.t]

category_data :: {String.t, String.t}

doc :: String.t

query :: String.t

t :: %CourseData{authors: authors, average_rating: float, category: String.t, enrolled: non_neg_integer, price: non_neg_integer, rating: non_neg_integer, source: String.t, subcategory: String.t, topic: String.t, url: String.t}

Functions

__struct__()

Defines the fixed structure of a course.

build_course_struct(arg1)

Specs

build_course_struct(response) :: CourseData.t when response: CoursesScraper.DocumentFetcher.response

Builds the structure of an individual Udemy course. It parses the data from an HTTP response and scrape the body to extract the needed data.

get_authors(doc, query)

Specs

get_authors(doc, query) :: authors

Get the list of authors from the HTML document of the course's website, based on a CSS query. This time the build_author_struct function from the CourseAuhor module is used to take care of the specific author data.

get_average_rating(doc, query)

Specs

get_average_rating(doc, query) :: avg_rating when avg_rating: float

Get the average rating from the HTML document of the course's website, based on a CSS query

get_category(arg)

Specs

get_category(category_data) :: category when category: String.t

get_category(category_data) :: subcategory when subcategory: String.t

get_category_data(doc, query)

Specs

get_category_data(doc, query) :: category_data

Get the course category from the HTML document of the course's website, based on a CSS query, and pattern match against the result to return the proper value.

get_enrolled(doc, query)

Specs

get_enrolled(doc, query) :: enrolled when enrolled: non_neg_integer

Get the enrolled information subtree from the course's HTML document and perform a Regex.scan against a fixed regular expression, in order to find the number of enrolled students.

get_price(doc, query)

Specs

get_price(doc, query) :: price when price: integer

Get the price from the HTML document of the course's website, based on a CSS query. This time the price was present in the head of the document as a metada field, so I applied some extra steps of transformations to get back the proper result.

get_rating(doc, query)

Specs

get_rating(doc, query) :: rating when rating: non_neg_integer

Get the rating (amount of students who rated the course) from the HTML document of the course's website, based on a CSS query

get_subcategory(arg)

get_topic(doc, query)

Specs

get_topic(doc, query) :: topic when topic: String.t

Get the course topic from the HTML document of the course's website, based on a CSS query.