Courses Scrapper v0.0.1 CourseData

Provides a set of functions to crawl the data from a course's HTML document. It brings the ability to build a struct with the needed information about a course.

Example

iex> {:ok, %HTTPoison.Response{status_code: 200, body: body}} = HTTPoison.get("https://www.udemy.com/android-basico/")
{:ok, %HTTPoison.Response{status_code: 200, body: "..."}}
iex> CourseData.extract_data({:ok, body, "android-basico"})
%CourseData{
    url: "https://udemy.com/android-basico",
    topic: "Android básico",
    subcategory: "Mobile Apps",
    source: "Udemy",
    rating: 77,
    price: 20,
    enrolled: 4278,
    category: "Development",
    average_rating: 4.6,
    authors: [
        %CourseAuthor{
            youtube: "https://www.youtube.com/josedlujan1",
            twitter: "https://twitter.com/josedlujan",
            name: "José Luján",
            linkedin: "https://linkedin.com/in/josedlujan/",
            googleplus: "none",
            facebook: "https://www.facebook.com/josedlujan"
        }
    ]
}

Functions

__struct__()

Defines the fixed structure of a course.

build_course_struct(arg1)

Builds the structure of an individual Udemy course. It parses the data from a HTTP response and scrape the body to extract the needed data.

get_authors(doc, query)

Get the list of authors from the HTML document of the course's website, based on a CSS query. This time the build_author_struct function from the CourseAuhor module is used to take care of the specific author data.

get_average_rating(doc, query)

Get the average rating from the HTML document of the course's website, based on a CSS query and pattern match against the result to return the proper value

get_category(doc, query)

Get the course category from the HTML document of the course's website, based on a CSS query, and pattern match against the result to return the proper value.

get_enrolled(doc, query)

Get the enrolled information subtree from the course's HTML document and perform a Regex.scan against a fixed regular expression, in order to find the rating and the number of enrolled students as a list.

get_price(doc, query)

Get the price from the HTML document of the course's website, based on a CSS query. This time the price was present in the head of the document as a metada field, so I applied some extra steps of transformations to get back the proper result.

get_topic(doc, query)

Get the course topic from the HTML document of the course's website, based on a CSS query.