Courses Scrapper v0.0.1 CoursesScraper.DocumentFetcher

provides a set of high level functions to fetch the data from a website and handle the result into a meaningful value.

Functions

build_course_url(course_path)

Build the full URL to reach the course's website

fetch(course_path)

Fetch the data of an Udemy course's website using the HTTPoison library to simulate an HTTP client, and handle the result to return some meaningful data to the caller of the function. This function is embedded into a try do - rescue block because the HTTPoison library has a bug that limits a client to make "too many" requests.

handle_response(arg1, path)

Handle the response that comes from an HTTP GET request

Handle a succesfull response with the form of {:ok, %HTTPoison.Response{status_code: 200, body: body}, transparently log a message to the console and return a tuple with an atom indicating the success, the body of the response and the path requested

Handle an useless response and return some meaningful data to the caller.

Handle an erroneous response and return some meaningful data to the caller.